# back-end/swagger.json
{
    "swagger": "3.0",
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "AgileOne CRUD",
        "description": "Recruiting Application",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:2000/api/",
            "description": "Dev"
        }
    ],
    "components": {        
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "schemes": ["http","https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/admin/get_singleuser": {
            "post": {
                "tags": ["Admin"],
                "description": "Get Single User",
                "operationId": "get_singleuser",
                "summary": "Get Single User Details in system",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "loginType": "admin/employee",
                                "user_id": "User ID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/admin/updatePassword": {
            "post": {
                "tags": ["Admin"],
                "description": "Update Password",
                "operationId": "updatePassword",
                "summary": "Update Password Details in system",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "new_password": "New Password",
                                "confirm_password": "Confirm New Password",
                                "user_id": "User ID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/updateProfile": {
            "post": {
                "tags": ["Admin"],
                "description": "Update User Profile",
                "operationId": "updateProfile",
                "summary": "Update User Profile Details in system",
                "consumes": "multipart/form-data",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "role": { "type": "integer" },
                                    "file": { "type": "string", "format": "binary" },
                                    "admin_name": { "type": "string" },
                                    "email": { "type": "string" },
                                    "mobile": { "type": "string" },
                                    "designation": { "type": "string" },
                                    "old_picture": { "type": "string" },
                                    "user_id": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/deleteUser": {
            "delete": {
                "tags": ["Admin"],
                "description": "Delete User",
                "operationId": "deleteUser",
                "summary": "Delete User Details from system",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "user_id": "User ID"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/updateUser": {
            "post": {
                "tags": ["Admin"],
                "description": "Update User Details",
                "operationId": "updateUser",
                "summary": "Update User Details in system",
                "consumes": "application/json",
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#"
                            },
                            "example": {
                                "role": "Role ID from get_roles api",
                                "admin_name": "User Name",
                                "email": "User Email",
                                "mobile": "User Mobile",
                                "designation": "User Designation",
                                "user_id": "User ID",
                                "updated_by": "Updated User ID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}